requirements/base.txt
aioredis==2.0.1
alembic==1.9.1
bcrypt
celery==5.2.7
redis==4.3.3
requests==2.31.0
databases[asyncpg]==0.7.0
orjson==3.8.4
psycopg2-binary==2.9.9
python-jose==3.3.0
SQLAlchemy==1.4.46
httpx==0.23.3
boto3==1.28.68
python-multipart==0.0.6
flower == 2.0.1
m3u8==3.6.0
pydantic[email]==1.10.4
fastapi==0.88.0
uvicorn[standard]==0.20.0

sentry-sdk==1.12.1

opentelemetry-api==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
prometheus-client==0.19.0
opentelemetry-instrumentation-celery==0.42b0